fix(playerbots): resolve auction feature build issues on AzerothCore branch APIs
switch auction house lookups to faction-template APIs used by current core
fix auction map retrieval call signature for buyer flow
resolve ObjectGuid parsing ambiguity in auction intent deserialization
eliminate quality narrowing warning in auction buy candidate creation
keep AH sell and buy behavior unchanged while restoring successful compilation and linking

+++ src/Ai/World/Rpg/AuctionIntents.cpp
@@ -78,7 +78,7 @@ namespace AuctionBot
 
             try
             {
-                ObjectGuid itemGuid(std::stoull(fields[0]));
+                ObjectGuid itemGuid(static_cast<uint64>(std::stoull(fields[0])));
                 if (!itemGuid.IsItem())
                     continue;
